requirements.txt
mindspore-ascend>=1.5.0
numpy
easydict
tqdm

